requirements.txt
# Minimum Django and REST framework version
Django<1.9
djangorestframework<3.3

# Test requirements
pytest-django==2.9.1
pytest==2.8.2
pytest-cov==1.6
flake8==2.2.2

# wheel for PyPI installs
wheel==0.24.0

# MkDocs for documentation previews/deploys
mkdocs==0.11.1
